+++ MOP/src/Vehicles/Cases/Kekmet.cs
@@ -56,6 +56,8 @@ namespace MOP.Vehicles.Cases
 
                 Position = transform.localPosition;
                 Rotation = transform.localRotation;
+
+                colliders.parent = temporaryParent;
             }
 
             gameObject.SetActive(enabled);
@@ -70,6 +72,8 @@ namespace MOP.Vehicles.Cases
             if (enabled)
             {
                 MoveNonDisableableObjects(null);
+                colliders.parent = transform;
+                colliders.localPosition = colliderPosition;
             }
         }